[Fix/#60]: scrap -> posts로 통합
