+++ app.js
@@ -14,7 +14,6 @@ app.listen(process.env.PORT, () => {
   console.log(`Server listening on ${process.env.PORT}...`);
 });
 
-const scrapRouter = require("./routes/scrapRouter");
 const authRouter = require("./routes/authRouter");
 const userRouter = require("./routes/usersRouter");
 const postRouter = require("./routes/postRouter");
@@ -24,4 +23,3 @@ app.use("/users", userRouter);
 app.use("/posts", postRouter);
 app.use("/comments", commentRouter);
 app.use("/oauth", authRouter);
-app.use("/scrap", scrapRouter);
