+++ services/postService.js
@@ -5,7 +5,6 @@ const { StatusCodes } = require("http-status-codes");
 const CustomError = require("../utils/CustomError");
 const { getUserInfo } = require("../utils/getUserInfo");
 const { updatePointsAndLevel } = require("../utils/updateLevel");
-const scrapQuery = require("../queries/scrapQuery");
 
 const writePost = async (writerId, categoryId, title, content, links) => {
   let values = [writerId, categoryId, title, content];
@@ -144,7 +143,7 @@ const getPostDetail = async (postId, userId) => {
     let isScrapped = null;
     if (userId) {
       const { count } = await conn
-        .query(scrapQuery.getScrapCount, [postId, userId])
+        .query(postQuery.getScrapCount, [postId, userId])
         .then((res) => res[0][0]);
 
       isScrapped = count > 0 ? true : false;
@@ -251,4 +250,84 @@ const updatePost = async (
   }
 };
 
-module.exports = { writePost, getPosts, getPostDetail, deletePost, updatePost };
+const scrap = async (userId, postId) => {
+  try {
+    const { count } = await conn
+      .query(postQuery.getScrapCount, [postId, userId])
+      .then((res) => res[0][0]);
+
+    if (count > 0) {
+      throw new CustomError(
+        StatusCodes.BAD_REQUEST,
+        "이미 스크랩된 게시글입니다."
+      );
+    }
+
+    const { affectedRows } = await conn
+      .query(postQuery.scrap, [postId, userId])
+      .then((res) => res[0]);
+
+    return {
+      isSuccess: affectedRows > 0,
+      message: affectedRows > 0 ? "스크랩 성공" : "스크랩 실패",
+    };
+  } catch (err) {
+    throw err;
+  }
+};
+
+const deleteScrap = async (userId, postId) => {
+  try {
+    const { affectedRows } = await conn
+      .query(postQuery.deleteScrap, [userId, postId])
+      .then((res) => res[0]);
+
+    return {
+      isSuccess: affectedRows > 0,
+      message: affectedRows > 0 ? "스크랩 삭제 성공" : "스크랩 삭제 실패",
+    };
+  } catch (err) {
+    throw err;
+  }
+};
+
+const getScrapList = async (userId) => {
+  try {
+    const scrapResult = await conn
+      .query(postQuery.getScrapList, userId)
+      .then((res) => res[0]);
+
+    if (scrapResult.length === 0) {
+      return {
+        isSuccess: true,
+        message: "스크랩 리스트 조회 성공",
+        result: [],
+      };
+    }
+
+    const scrapPostId = scrapResult.map((result) => result.post_id);
+
+    const postResult = await conn
+      .query(postQuery.getScrapPosts, [scrapPostId])
+      .then((res) => res[0]);
+
+    return {
+      isSuccess: true,
+      message: "스크랩 리스트 조회 성공",
+      result: postResult,
+    };
+  } catch (err) {
+    throw err;
+  }
+};
+
+module.exports = {
+  writePost,
+  getPosts,
+  getPostDetail,
+  deletePost,
+  updatePost,
+  scrap,
+  deleteScrap,
+  getScrapList,
+};
