+++ services/scrapService.js
@@ -33,47 +33,6 @@ const getScrapList = async (userId) => {
   }
 };
 
-const scrap = async (userId, postId) => {
-  try {
-    const { count } = await conn
-      .query(scrapQuery.getScrapCount, [postId, userId])
-      .then((res) => res[0][0]);
-
-    if (count > 0) {
-      throw new CustomError(
-        StatusCodes.BAD_REQUEST,
-        "이미 스크랩된 게시글입니다."
-      );
-    }
-
-    const { affectedRows } = await conn
-      .query(scrapQuery.scrap, [postId, userId])
-      .then((res) => res[0]);
-
-    return {
-      isSuccess: affectedRows > 0,
-      message: affectedRows > 0 ? "스크랩 성공" : "스크랩 실패",
-    };
-  } catch (err) {
-    throw err;
-  }
-};
-
-const deleteScrap = async (userId, postId) => {
-  try {
-    const { affectedRows } = await conn
-      .query(scrapQuery.deleteScrap, [userId, postId])
-      .then((res) => res[0]);
-
-    return {
-      isSuccess: affectedRows > 0,
-      message: affectedRows > 0 ? "스크랩 삭제 성공" : "스크랩 삭제 실패",
-    };
-  } catch (err) {
-    throw err;
-  }
-};
-
 module.exports = {
   getScrapList,
   scrap,
